requirements.txt
ipython
PyYAML==3.13
